[*] use double values throught painting

+++ TinyRaspiGui/include/plotWidget.h
@@ -16,16 +16,16 @@ public slots:
 public:
 	CPlotWidget(QWidget *parent = Q_NULLPTR);
 
-	void SetMax(int max);
-	void SetMin(int min);
+    void SetMax(double max);
+    void SetMin(double min);
 	
 	void paintEvent(QPaintEvent* e);
 	void resizeEvent(QResizeEvent* e);
 
 private:
-	int						m_plotWidth;
-	int						m_plotMax;
-	int						m_plotMin;
+    double                  m_plotWidth;
+    double                  m_plotMax;
+    double                  m_plotMin;
 	QList<QList<double>>	m_plotValues;
 	bool					m_showGrid;
 	bool					m_boldCenterline;
